Move CookieBanner integration to @app/legal/CookieBannerConsent.tsx
